requirements.txt
asgiref==3.8.1
astroid==3.3.8
colorama==0.4.6
coverage==7.6.9
dill==0.3.9
Django==5.1.3
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
factory_boy==3.3.1
Faker==33.1.0
graphviz==0.20.3
isort==5.13.2
mccabe==0.7.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.10.0
pylint==3.3.3
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.2
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
